refactor(server): modularize mock routes and add faker for product data
The mock routes for account, auth, and products have been moved into separate modules for better maintainability. Additionally, the @faker-js/faker package has been added to generate realistic mock product data.
